is-default-class storageclass local-path
So by default, you get a PVC filled from Rancher, that will create a PV
that persists for the life of the cluster, or the PV will go away when
its corresponding PVC is deleted if the retain policy is set to delete.

This annotation on the storage class ensures our PVCs don't have to know
about (or opt in to) the policy in order to just work.